Updated CompleteCurrentLevel() Method in GameMgr
It will now unlock the next level, if there is one.
It will now write and save level progress data, using the new JSON I/O scripts.

It will now load to the next level if there is one.
If there is no next level, the main menu will be loaded.